move spine color seting to character
