+++ nekoyume/Assets/_Scripts/Game/VFX/Skill/BuffVFX.cs
@@ -1,15 +1,19 @@
 using UnityEngine;
 using Nekoyume.Game.Character;
+using Nekoyume.Model.BattleStatus;
 
 namespace Nekoyume.Game.VFX.Skill
 {
     public class BuffVFX : VFX
     {
         [field: SerializeField]
-        public Actor Target { get; set; }
+        public Character.Character Target { get; set; }
 
         [field: SerializeField]
         public virtual bool IsPersisting { get; set; } = false;
+
+        [field: SerializeField]
+        public Model.Buff.Buff Buff { get; set; }
 
 #if UNITY_EDITOR
         [field: SerializeField]
